+++ Source/EventFlow.EventStores.EventStore/Properties/AssemblyInfo.cs
@@ -23,22 +23,7 @@
 //
 
 using System.Reflection;
-using System.Runtime.InteropServices;
 
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
 [assembly: AssemblyTitle("EventFlow.EventStores.EventStore")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("EventFlow.EventStores.EventStore")]
-[assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("8fe7f61e-82f8-4e88-8998-abee202078db")]
+[assembly: AssemblyDescription("Event Store event store for EventFlow. Download it from https://geteventstore.com/.")]